refactor: rename SpriteSvg.tsx to sprite-svg.tsx and change related imports
