+++ src/components/sprite-svg.tsx
