+++ src/pages/Home/index.tsx
@@ -1,9 +1,14 @@
+import PaymentCards from '@features/payment-card/components/payment-cards';
 import { FC } from 'react';
 
 interface HomeProps {}
 
 const Home: FC<HomeProps> = ({}) => {
-  return <></>;
+  return (
+    <>
+      <PaymentCards />
+    </>
+  );
 };
 
 export default Home;
